refactor(datahub-client): move utf8 encoding code to eventFormatter
