+++ metadata-integration/java/datahub-client/src/main/java/datahub/event/EventFormatter.java
@@ -7,6 +7,7 @@ import com.linkedin.data.ByteString;
 import com.linkedin.data.template.JacksonDataTemplateCodec;
 import com.linkedin.mxe.GenericAspect;
 import com.linkedin.mxe.MetadataChangeProposal;
+
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.nio.charset.StandardCharsets;
@@ -33,7 +34,8 @@ public class EventFormatter {
 
   @SneakyThrows(URISyntaxException.class)
   public MetadataChangeProposal convert(MetadataChangeProposalWrapper mcpw) throws IOException {
-    String serializedAspect = dataTemplateCodec.dataTemplateToString(mcpw.getAspect());
+    
+    String serializedAspect = StringEscapeUtils.escapeJava(dataTemplateCodec.dataTemplateToString(mcpw.getAspect()));
     MetadataChangeProposal mcp = new MetadataChangeProposal().setEntityType(mcpw.getEntityType())
         .setAspectName(mcpw.getAspectName())
         .setEntityUrn(Urn.createFromString(mcpw.getEntityUrn()))
